+++ src/cards/Scorers/soccer-live-scorers.js
@@ -51,7 +51,7 @@ class SoccerLiveScorersCard extends LitElement {
 
   render() {
     applySkin(this, this._config);
-    if (!this.hass || !this._config) return renderLoading('Loading...');
+    if (!this.hass || !this._config) return renderLoading(this._t('ui.loading'));
     const entityId = this._config.entity;
     const stateObj = this.hass.states[entityId];